GIFPT_AI/requirements.txt
Django==4.2.26
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
openai==1.40.6
requests==2.32.3
httpx==0.27.2
gunicorn>=21.2
PyMuPDF==1.24.9
Pillow==10.3.0
boto3>=1.35.0
manim==0.19.0
jsonschema==4.21.1
jinja2==3.1.4
pydantic>=2.7,<3
langsmith>=0.1.0,<1
